amends packages sorted + spec refactors
