+++ test/forms/amends/AgentsFormProviderSpec.scala
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package forms
+package forms.amends
 
+import forms.AgentsFormProvider
 import forms.behaviours.StringFieldBehaviours
 import play.api.data.FormError
