+++ test/forms/amends/AmendDirectExportPlasticPackagingFormProviderSpec.scala
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package forms
+package forms.amends
 
+import forms.AmendDirectExportPlasticPackagingFormProvider
 import forms.behaviours.IntFieldBehaviours
 import play.api.data.FormError
