+++ test/navigation/FakeNavigator.scala
@@ -20,7 +20,7 @@ import play.api.mvc.Call
 import pages._
 import models.{Mode, UserAnswers}
 
-class FakeNavigator(desiredRoute: Call) extends Navigator {
+class FakeNavigator(desiredRoute: Call) extends Navigator(new AmendsJourneyNavigator, returns = new ReturnsJourneyNavigator) {
 
   override def nextPage(page: Page, mode: Mode, userAnswers: UserAnswers): Call =
     desiredRoute
